Relax true-on-policy overrides and preserve raw rollout scores
- Keep rollout log probabilities in FP32 for all training precisions so raw
  server scores survive ingestion and BF16 rounding cannot conceal decode
  mismatches; retain the former precision selectors as commented references.
- Use the TOP LM head's existing BF16 logits without an unconditional cast,
  while retaining the BF16 conversion after non-unit temperature scaling.
- Stop forcing prefill log-probability recomputation, batch-invariant mode,
  global deterministic mode, FP32 residual settings, and CPU initialization
  during TOP validation; preserve caller choices and the required local
  transformer, sequence-parallel, fusion, and Qwen3 contract settings.
- Continue deriving SGLang deterministic inference and FA3 attention from TOP,
  but preserve the caller's prefill-only deterministic inference setting to
  avoid forcing a mode that breaks graph capture on the tested A100 runtime.
- Simplify the Odyssey TOP launcher by commenting redundant parity flags and
  legacy log-probability/RMSNorm selectors; keep explicit deterministic mode
  and NCCL/NVTE/cuBLAS environment settings for non-TOP launches.
- Set Odyssey debug defaults to round-robin routing, micro and rollout batch
  sizes of two, disabled W&B, and a bitwise-parity-debug experiment prefix;
  remove the effective default task-family quota assignment.
- Default evaluation to FlashInfer CUDA normalization alongside the Torch
  fallback, covering newer sgl_kernel calls that bypass the Torch selector
  while retaining environment overrides and CUDA graph support.
- Extend argument validation coverage for optional TOP settings, assert FP32
  storage preserves a real decode score, update SGLang prefill expectations,
  and mark the TOP test module as CPU-only with a direct pytest entry point.
- Validation: all 37 tests in test_megatron_argument_validation.py and
  test_true_on_policy.py passed; both modified launchers passed bash -n,
  and git diff --check reported no whitespace errors.

+++ slime/backends/megatron_utils/loss.py
@@ -528,9 +528,9 @@ def get_log_probs_and_entropy(
     logits = logits.contiguous()
     true_on_policy = getattr(args, "true_on_policy_mode", False)
     if true_on_policy:
-        # Match SGLang Sampler exactly: cast logits before temperature division,
-        # then retain BF16 for the batch-invariant log-softmax kernel.
-        logits = logits.bfloat16()
+        # The TOP LM head already supplies BF16 logits. Retain BF16 after any
+        # temperature division for the batch-invariant log-softmax kernel.
+        # logits = logits.bfloat16()  # The TOP LM head already returns BF16; no_logits_cast kept exact parity.
         if rollout_temperature != 1.0:
             logits = logits.div(rollout_temperature).bfloat16()
     T = logits.size(0)
